+++ crates/desktop/src/lib.rs
@@ -67,16 +67,6 @@ pub fn run() {
       is_maximized,
       launch_game
     ])
-    .setup(|app| {
-      if cfg!(debug_assertions) {
-        app.handle().plugin(
-          tauri_plugin_log::Builder::default()
-            .level(log::LevelFilter::Info)
-            .build(),
-        )?;
-      }
-      Ok(())
-    })
     .run(tauri::generate_context!())
     .expect("error while running tauri application");
 }